feat：同步社区最新版本，增加和修改 build 配置，重新生成 build 产物 (#33)

+++ lib/typescript/specs/NativeBlurhashModule.d.ts
@@ -1,8 +1,7 @@
-/// <reference types="react-native/types/modules/codegen" />
 import { type TurboModule } from 'react-native';
-import type { Int32 } from 'react-native/Libraries/Types/CodegenTypes';
+import type { Double } from 'react-native/Libraries/Types/CodegenTypes';
 export interface Spec extends TurboModule {
-    createBlurhashFromImage: (imageUri: string, componentsX: Int32, componentsY: Int32) => Promise<string>;
+    createBlurhashFromImage: (imageUri: string, componentsX: Double, componentsY: Double) => Promise<string>;
     clearCosineCache: () => void;
 }
 declare const _default: Spec;
